chore: refactor to custom hook
